Move global metadata DB sources into the clp namespace.
