+++ components/core/src/clp/clg/CMakeLists.txt
@@ -1,5 +1,12 @@
 set(
         CLG_SOURCES
+        ../GlobalMetadataDB.hpp
+        ../GlobalMetadataDBConfig.cpp
+        ../GlobalMetadataDBConfig.hpp
+        ../GlobalMySQLMetadataDB.cpp
+        ../GlobalMySQLMetadataDB.hpp
+        ../GlobalSQLiteMetadataDB.cpp
+        ../GlobalSQLiteMetadataDB.hpp
         "${PROJECT_SOURCE_DIR}/src/BufferReader.cpp"
         "${PROJECT_SOURCE_DIR}/src/BufferReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/database_utils.cpp"
@@ -22,13 +29,6 @@ set(
         "${PROJECT_SOURCE_DIR}/src/FileReader.hpp"
         "${PROJECT_SOURCE_DIR}/src/FileWriter.cpp"
         "${PROJECT_SOURCE_DIR}/src/FileWriter.hpp"
-        "${PROJECT_SOURCE_DIR}/src/GlobalMetadataDB.hpp"
-        "${PROJECT_SOURCE_DIR}/src/GlobalMetadataDBConfig.cpp"
-        "${PROJECT_SOURCE_DIR}/src/GlobalMetadataDBConfig.hpp"
-        "${PROJECT_SOURCE_DIR}/src/GlobalMySQLMetadataDB.cpp"
-        "${PROJECT_SOURCE_DIR}/src/GlobalMySQLMetadataDB.hpp"
-        "${PROJECT_SOURCE_DIR}/src/GlobalSQLiteMetadataDB.cpp"
-        "${PROJECT_SOURCE_DIR}/src/GlobalSQLiteMetadataDB.hpp"
         "${PROJECT_SOURCE_DIR}/src/Grep.cpp"
         "${PROJECT_SOURCE_DIR}/src/Grep.hpp"
         "${PROJECT_SOURCE_DIR}/src/ir/LogEvent.hpp"
