+++ components/core/src/clp/clp/compression.cpp
@@ -6,12 +6,12 @@
 #include <boost/filesystem/operations.hpp>
 #include <boost/uuid/random_generator.hpp>
 
-#include "../../GlobalMySQLMetadataDB.hpp"
-#include "../../GlobalSQLiteMetadataDB.hpp"
 #include "../../spdlog_with_specializations.hpp"
 #include "../../streaming_archive/writer/Archive.hpp"
 #include "../../streaming_archive/writer/utils.hpp"
 #include "../../Utils.hpp"
+#include "../GlobalMySQLMetadataDB.hpp"
+#include "../GlobalSQLiteMetadataDB.hpp"
 #include "FileCompressor.hpp"
 #include "utils.hpp"
