+++ components/core/src/clp/clp/decompression.cpp
@@ -7,12 +7,12 @@
 
 #include "../../ErrorCode.hpp"
 #include "../../FileWriter.hpp"
-#include "../../GlobalMySQLMetadataDB.hpp"
-#include "../../GlobalSQLiteMetadataDB.hpp"
 #include "../../spdlog_with_specializations.hpp"
 #include "../../streaming_archive/reader/Archive.hpp"
 #include "../../TraceableException.hpp"
 #include "../../Utils.hpp"
+#include "../GlobalMySQLMetadataDB.hpp"
+#include "../GlobalSQLiteMetadataDB.hpp"
 #include "FileDecompressor.hpp"
 
 using std::cerr;
